Move create_disassembler/assembler to core_utils and refactor compile
